requirements.txt
fastapi
uvicorn
langchain
requests
pydantic
python-dotenv
faiss-cpu
langchain 
langchain-google-genai
langchain-community
sentence-transformers 